AJAX: 'Infinity Scroll' add. Now items (users, microposts) can be loaded by parts.
