+++ app/assets/javascripts/application.js
@@ -27,4 +27,22 @@ document.addEventListener("turbolinks:load", function() {
             next: "fa fa-arrow-right"
         }
     });
+
+   $("div[id|='reply-to-micropost']").hide()
+   window.toggleReply = function (id) {
+        $("#reply-to-micropost-" + id.toString()).slideToggle(500);
+    }
+
+    if ($('#scrolling').size() > 0) {
+        $(window).on('scroll', function() {
+            var url;
+            url = $('.pagination .next_page a').attr('href');
+            if (url && $(window).scrollTop() > $(document).height() - $(window).height() - 50) {
+                document.getElementById("loading").style.display = 'block';
+                $('.pagination').html('<h4>Loading...</h4>');
+                $.getScript(url);
+            }
+        });
+        return;
+    }
 })
