+++ app/assets/stylesheets/likes.scss
@@ -30,3 +30,20 @@
   background: linear-gradient(rgba(40, 84, 118, 0.5), rgba(99, 158, 202, 1) );
 }
 
+.like-form {
+  padding: 10px 5px;
+}
+
+div[id|="liker-avatars"] {
+  margin: 0;
+  padding-left: 10px;
+  .avatar {
+    margin: 2px 2px;
+  }
+  a {
+    padding: 0;
+  }
+}
+
+
+
