+++ app/assets/stylesheets/microposts.scss
@@ -32,7 +32,7 @@
   .timestamp {
     color: darken(#e7e7e7, 20%);
     display: block;
-    margin-left: 105px;
+    margin-left: 125px;
     padding: 10px;
     a {
       color: darken(#b3e0ff, 50%);
@@ -66,4 +66,18 @@
   .container {
     margin-left: 0;
   }
+}
+
+button[id|="show-replies"] {
+  font-family: 'Oswald', sans-serif;
+  color: darken(#b3e0ff, 50%);
+  width: 100%;
+  border: none;
+  background: none;
+  outline: none;
+  padding-bottom: 5px;
+}
+
+button[id|="show-replies"]:focus {
+  outline: none;
 }
