+++ app/assets/stylesheets/users.scss
@@ -42,3 +42,7 @@
     }
   }
 }
+
+.pagination {
+  display: none;
+}
