+++ app/controllers/tags_controller.rb
@@ -2,6 +2,8 @@ class TagsController < ApplicationController
 
   def show
     @tag =  ActsAsTaggableOn::Tag.find(params[:id])
-    @micropost = Micropost.tagged_with(@tag.name)
+    @tagged_microposts = Micropost.tagged_with(@tag.name)
+    #@tagged_microposts << Reply.tagged_with(@tag.name)
+    @reply_tags = Reply.tagged_with(@tag.name)
   end
 end
