+++ app/controllers/users_controller.rb
@@ -4,7 +4,7 @@ class UsersController < ApplicationController
   before_action :admin_user,     only: :destroy
 
   def index
-    @users = User.reorder('email').paginate(page: params[:page], :per_page => 4)
+    @users = User.reorder('email').paginate(page: params[:page], per_page: 6)
   end
 
   def show
@@ -68,24 +68,24 @@ class UsersController < ApplicationController
   def following
     @title = 'Following'
     @user = User.find(params[:id])
-    @users = @user.following.paginate(page: params[:page], :per_page => 30)
+    @users = @user.following
     render 'show_follow'
   end
 
   def followers
     @title = 'Followers'
     @user = User.find(params[:id])
-    @users = @user.followers.paginate(page: params[:page], per_page: 30)
+    @users = @user.followers
     render 'show_follow'
   end
 
   def liked_posts
     @user = User.find(params[:id])
-    @liked_posts = @user.liked_microposts.paginate(page: params[:page], per_page: 10)
+    @liked_posts = @user.liked_microposts.paginate(page: params[:page], per_page: 5)
   end
 
   def show_microposts
-    @microposts = @user.microposts.paginate(page: params[:page])
+    @microposts = @user.microposts.paginate(page: params[:page], per_page: 5)
     @micropost = current_user.microposts.build if logged_in?
   end
