+++ app/views/landing_pages/home.js.erb
@@ -1,0 +1,10 @@
+$('.microposts').append('<%= j render(@feed_items) %>');
+$("div[id|='reply-to-micropost']").hide()
+<% if @feed_items.next_page %>
+  $('.pagination').replaceWith('<%= j will_paginate(@feed_items) %>');
+  $('#loading').hide();
+<% else %>
+  $('.pagination').remove();
+  $('#loading').remove();
+<% end %>
+<% sleep 0.5 %>
