+++ app/views/microposts/_feed.html.erb
@@ -6,6 +6,9 @@
     <%= render @feed_items %>
   </ol>
   <% if logged_in? %>
-    <%= will_paginate @feed_items %>
+    <div id="scrolling">
+      <div id="loading"><img src="/ajax-loader.gif" alt="Loading..."></div>
+      <%= will_paginate @feed_items %>
+    </div>
   <% end %>
 <% end %>
