+++ app/views/microposts/_micropost.html.erb
@@ -9,16 +9,18 @@
       <% end %>
       <span class="content"><%= process_content(micropost) %></span>
       <div class="post-image"><%= image_tag micropost.image.url if micropost.image? %></div>
-      <div class="like-panel">
-        <%= render 'microposts/like_form', micropost: micropost %>
-        <%= render 'microposts/likes', micropost: micropost  %>
-        <%= render 'microposts/post_likers', micropost: micropost %>
+      <div class="like-panel row">
+        <div class="like-form col-sm-2">
+          <%= render 'microposts/like_form', micropost: micropost %>
+          <%= render 'microposts/likes', micropost: micropost  %>
+        </div>
+        <div><%= render 'microposts/post_likers', micropost: micropost %> </div>
       </div >
       <span class="timestamp">Posted <%= time_ago_in_words(micropost.created_at) %> ago.
         <span class="remove-post">
         <% if current_user?(micropost.user) %>
           <%= link_to micropost, method: :delete,
-                                data: { confirm: "Remove your post?" },
+                                data: { confirm: 'Remove your post?' },
                                 class: 'post-delete-link' do %>
           <i class="fa fa-lg fa-trash"></i>
           <% end %>
@@ -26,12 +28,15 @@
         </span>
       </span>
       <% if logged_in? %>
-        <hr>
-        <div class="replies-panel">
-          <%= render 'replies/reply', micropost: micropost %>
+        <hr><button onclick="toggleReply(<%= micropost.id %>)" id="show-replies-<%= micropost.id %>" class="toggle"><i class="fa fa-reply" aria-hidden="true"></i> REPLY</button>
+        <div id="reply-to-micropost-<%= micropost.id %>">
+          <hr>
+          <div class="replies-panel">
+            <%= render 'replies/reply', micropost: micropost %>
+          </div>
+          <hr>
+          <%= render 'replies/reply_form', micropost: micropost %>
         </div>
-        <hr>
-        <%= render 'replies/reply_form', micropost: micropost %>
       <% end %>
     </li>
   </div>
