+++ app/views/replies/_reply.html.erb
@@ -4,7 +4,7 @@
       <li id="reply-<%= reply.id %>">
         <%= link_to avatar_for(reply.user, 65), reply.user %>
         <h4 class="reply-owner"><%= link_to reply.user.email , reply.user %></h4>
-        <span class="reply-content"><%= reply.content %></span>
+        <span class="reply-content"><%= process_content(reply) %></span>
         <div class="reply-image"><%= image_tag reply.image.url if reply.image? %></div>
         <span class="timestamp">
           Reply <%= time_ago_in_words(reply.created_at) %> ago.
@@ -16,9 +16,9 @@
             <i class="fa fa-lg fa-trash"></i>
           <% end %>
         <% end %>
-       </span>
-      </span>
-    </li>
+        </span>
+        </span>
+      </li>
     <% end %>
   <% else %>
     <div class="empty-zone">
