+++ app/views/replies/create.js.erb
@@ -1,2 +1,2 @@
-$("#replies-<%= @micropost.id %>").hide().fadeIn(1000).html("<%= escape_javascript(render 'replies/reply', micropost: @micropost)%>");
+$("#replies-<%= @micropost.id %>").hide().fadeIn(500).html("<%= escape_javascript(render 'replies/reply', micropost: @micropost)%>");
 $("#reply-textarea-<%= @micropost.id %>").val('')
