+++ app/views/replies/destroy.js.erb
@@ -1,1 +1,1 @@
-$("#reply-<%= @reply.id %>").slideUp(1000, function(){$(this).remove()});
+$("#reply-<%= @reply.id %>").fadeOut(500, function(){$(this).remove()});
