+++ app/views/users/index.html.erb
@@ -1,8 +1,9 @@
 <% provide(:title, 'All users') %>
 <h1>All users</h1>
-
-<%= will_paginate %>
-  <ul class="users">
+  <div class="users">
     <%= render @users.sort { |a,b| a.email.downcase <=> b.email.downcase } %>
-  </ul>
-<%= will_paginate %>
+  </div>
+  <div id="scrolling">
+    <div id="loading"><img src="/ajax-loader.gif" alt="Loading..."></div>
+    <%= will_paginate  @users %>
+  </div>
