+++ app/views/users/liked_posts.js.erb
@@ -1,0 +1,10 @@
+$('.microposts').append('<%= j render(@liked_posts) %>');
+$("div[id|='reply-to-micropost']").hide()
+<% if @liked_posts.next_page %>
+  $('.pagination').replaceWith('<%= j will_paginate(@liked_posts) %>');
+  $('#loading').hide();
+<% else %>
+  $('.pagination').remove();
+  $('#loading').remove();
+<% end %>
+<% sleep 1 %>
