+++ app/views/users/show.html.erb
@@ -27,7 +27,10 @@
       <ol class="microposts">
         <%= render @microposts %>
       </ol>
-      <%= will_paginate @microposts %>
+      <div id="scrolling">
+        <div id="loading"><img src="/ajax-loader.gif" alt="Loading..."></div>
+        <%= will_paginate @microposts %>
+      </div>
     <% else %>
       <div class="empty-zone">
         <hr>
